+++ changelog.md
@@ -1,1 +1,2 @@
-- fixed an issue with configs
+- blit modifiers can target a texture ID or a regex.
+- configs have feedback when you change a value. they were also moved in the config folder